My logo and banner submission
Took the original from nvllsm found at
https://nextcloud.nullsum.net/index.php/s/cBxQCy6CnakRpfR
And made a 512x512 square for the logo. I then did my best to center the logo.

For the banner i added 1024 to the right to make a rectangle and used the bundled chewy font to make the text. The "y" lines up with the bottom of the outer triangle and the "J" lines up with the bottom on the inner. I then shifted it all right to create proper and more even margins.